requirements.txt
voyageai
pillow>=9.0.0
pymongo>=4.0.0
python-dotenv>=1.0.0
